+++ examples/apps/image-folder-support/README.md
@@ -1,0 +1,22 @@
+## Introduction
+
+This example is a streamlit app that supports running inference with a local image folder.
+
+In particular this app allows a user to run inference with a local image folder, visualize the prediction result, analyze the prediction distribution, and export the prediction result to an CSV file.
+
+### Target Audience
+
+1. Basic Python programming skill.
+2. Basic understanding of the streamlit library. See [here](https://docs.streamlit.io/library/get-started/main-concepts) for more information.
+
+### Installation
+
+```
+pip install -r examples/apps/image-folder-support/requirements.txt
+```
+
+### Launch the App
+
+```
+streamlit run examples/apps/image-folder-support/app.py
+```
