+++ examples/apps/image-folder-support/requirements.txt
